requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
numpy==1.23.5
tensorflow==2.12.0
Pillow==10.0.0
waitress==2.1.2